requirements.txt
pillow == 3.4.1
pycorenlp == 0.3.0
spacy == 0.101.0
beautifulsoup4 == 4.5.1
